Add global reordering of mesh and checksum
The global reordering of the mesh is needed to control the order of the
mesh for runs without parallelization in domains. Moreover, it can now
be controled separately from the local mesh order.

The checksum enables restarting from meshes with the same ordering
without further steps and with different global index orderings with a
reordering.

+++ src/math/checksum_interface.F90
@@ -26,7 +26,7 @@ module checksum_interface_oct_m
       implicit none
       integer,    intent(in)  :: algorithm
       integer(8), intent(in)  :: narray
-      integer,    intent(in)  :: array
+      integer(8), intent(in)  :: array
       integer(8), intent(out) :: checksum
     end subroutine checksum_calculate
   end interface
